requirements.txt
streamlit==1.33.0
python-dotenv==1.0.1
langchain==0.1.16
langchain-groq==0.1.3
langchain-community==0.0.34
chromadb==0.4.24
sentence-transformers==2.7.0
pypdf==4.2.0
wikipedia==1.4.0
requests==2.31.0